Use nullable for CLI.

+++ cli/Squidex.CLI/Squidex.CLI/Commands/Implementation/FileSystem/Default/DefaultFile.cs
@@ -5,6 +5,7 @@
 //  All rights reserved. Licensed under the MIT license.
 // ==========================================================================
 
+using System;
 using System.IO;
 
 namespace Squidex.CLI.Commands.Implementation.FileSystem.Default
@@ -21,6 +22,8 @@ namespace Squidex.CLI.Commands.Implementation.FileSystem.Default
 
         public bool Exists => fileInfo.Exists;
 
+        public bool Readonly { get; init; }
+
         public DefaultFile(FileInfo fileInfo, string fullLocalName)
         {
             this.fileInfo = fileInfo;
@@ -28,19 +31,29 @@ namespace Squidex.CLI.Commands.Implementation.FileSystem.Default
             FullLocalName = fullLocalName;
         }
 
-        public void Delete()
-        {
-            fileInfo.Delete();
-        }
-
         public Stream OpenRead()
         {
             return fileInfo.OpenRead();
         }
 
+        public void Delete()
+        {
+            if (Readonly)
+            {
+                throw new InvalidOperationException("The file system is in readonly mode.");
+            }
+
+            fileInfo.Delete();
+        }
+
         public Stream OpenWrite()
         {
-            Directory.CreateDirectory(fileInfo.Directory.FullName);
+            if (Readonly)
+            {
+                throw new InvalidOperationException("The file system is in readonly mode.");
+            }
+
+            Directory.CreateDirectory(fileInfo.Directory!.FullName);
 
             return new FileStream(fileInfo.FullName, FileMode.Create, FileAccess.Write);
         }
